+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Support `application/flowfile-v1` contentType in ingress
 
 ### Changed
+- Replaced `resetFromSnapshot` with `importSnapshot`
 
 ### Deprecated